+++ mlxtend/preprocessing/dense_transformer.py
@@ -7,21 +7,28 @@
 # License: BSD 3 clause
 
 
-class DenseTransformer(object):
+from sklearn.base import BaseEstimator
+from scipy.sparse import issparse
 
-    """Convert a sparse matrix into a dense matrix."""
 
-    def __init__(self, some_param=True):
-        pass
+class DenseTransformer(BaseEstimator):
+    """Convert a sparse array into a dense array."""
+
+    def __init__(self, return_copy=True):
+        self.return_copy = return_copy
+        self.is_fitted = False
 
     def transform(self, X, y=None):
-        return X.toarray()
+        if issparse(X):
+            return X.toarray()
+        elif self.return_copy:
+            return X.copy()
+        else:
+            return X
 
     def fit(self, X, y=None):
+        self.is_fitted = True
         return self
 
     def fit_transform(self, X, y=None):
-        return X.toarray()
-
-    def get_params(self, deep=True):
-        return {'some_param': True}
+        return self.transform(X=X, y=y)